# Optimize __str_base10()