@visitorckw 2024-08-10
I have updated the commit message to make the process of how I tested and obtained the experimental data clearer. For each test range, values within the range were used as inputs, and __str_base10() was called 100,000,000 times with these inputs uniformly distributed. The execution time was measured using the time command. The test code is as follows:

```
int x = 0, i;
const int range = 1000;
char s[INT_BUF_LEN];

for (i = 0; i < 100000000; i++) {
    __str_base10(s, x);
    x = (x + 1) % range;
}
```