@DrXiao 2024-08-10
@visitorckw , I apologize that I forgot to read your commit messages first. But I have understood that the timing data is collected by your test code and `time` command. 
Thank you for your clarification!