@visitorckw 2024-08-10
FWIW, I wrote a unit test to compare the results of the new function with the old one to verify correctness. Within the range [0, INT_MAX], the results of both functions are identical.

Testing code:
```
int x = 0;
char s1[INT_BUF_LEN];
char s2[INT_BUF_LEN];

while (1) {
    __str_base10(s1, x);
    new__str_base10(s2, x);
    if (strcmp(s1, s2)) {
        printf("Wrong answer when x = %d\n", x);
        return 0;
    }
    if (x == INT_MAX) break;
    x++;
}
```